Add files via upload
Firstly we imported modules like datetime, random, string, and pymysql.The datetime module is used to manage date and time information. Random and string modules are used to generate random ticket numbers. The pymysql module is used to connect to and communicate with a MySQL database.
Next, the class 'Vehicle' is defined, which takes in six arguments - vehicle_number, vehicle_type, contact_number, ticket_number, entry_time, and exit_time. The constructor initializes the class attributes with the corresponding values.
The class also includes two methods - 'get_duration'and 'get_cost'. The 'get_duration' method calculates the duration of the parking and returns it. The 'get_cost' method calculates the parking cost based on the duration and returns it.
The 'print_receipt' method prints the parking receipt containing vehicle details, entry and exit time, duration, and the total cost.
After the class definition, the code connects to a MySQL database using pymysql and prompts the user to enter vehicle details such as vehicle number, type, contact number, entry, and exit time.
It generates a random ticket number using the string and random modules and creates a Vehicle object using the user-entered details. It then calls the 'print_receipt' method to display the parking receipt.
The code then inserts the parking record into the MySQL database using an SQL query and commits the changes. Finally, the connection to the database is closed.